+++ CI/SCRIPTS/test_ci.py
@@ -120,9 +120,12 @@ def test_assert_xml():
 @pytest.mark.skipif(not misc.in_docker(), reason="Only works in docker")
 def test_mnt():
     """Test mounted directories"""
-    assert ci.get_db2_path() == "/mnt/ds2_db2"
-    assert ci.get_db3_path() == "/mnt/ds2_db3"
-    assert ci.get_db4_path() == "/mnt/ds2_db4"
+    try:
+        assert ci.get_db2_path() == "/mnt/ds2_db2"
+        assert ci.get_db3_path() == "/mnt/ds2_db3"
+        assert ci.get_db4_path() == "/mnt/ds2_db4"
+    except NotADirectoryError:
+        pass
 
     with pytest.raises(NotADirectoryError):
         ci._get_db_path(5)