+++ sertit/__init__.py
@@ -20,4 +20,4 @@
 .. include:: ../README.md
 """
 
-__version__ = "1.5.0"
+__version__ = "1.6.0"